pdv/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-dotenv>=1.0.0
requests>=2.31.0
jinja2>=3.1.0
pywebview>=5.0.0
pyinstaller>=6.0.0
